+++ lofigirl_server/Cargo.toml
@@ -11,9 +11,11 @@ lofigirl_sys = { path = "../lofigirl_sys"}
 lofigirl_shared_common = { path = "../lofigirl_shared_common"}
 lofigirl_shared_listen = { path = "../lofigirl_shared_listen"}
 serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
 actix-web = "4.8"
 actix-cors = "0.7"
 actix-rt = "2.10"
+actix-ws = "0.3"
 sqlx = { version = "0.8", features = ["sqlite", "runtime-tokio-native-tls"] }
 uuid = { version = "1.8", features = ["v4"] }
 thiserror = "1.0"
@@ -24,3 +26,5 @@ url = "2.5"
 parking_lot = "0.12"
 tracing = "0.1"
 tracing-subscriber = "0.3"
+tokio = { version = "1.39", features = ["sync"] }
+futures-util = "0.3"
